scripts/requirements.txt
superinvoke==1.0.5
python-dotenv==1.0.0
pyserial==3.5
matplotlib==3.8.1